lineage-mcp-tray/pyproject.toml
[project]
name = "lineage-mcp-tray"
version = "0.1.0"
description = "System tray companion for lineage-mcp"
requires-python = ">=3.10"
dependencies = [
    "pystray>=0.19",
    "Pillow>=9.0",
]

[project.scripts]
lineage-tray = "lineage_tray.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]
